% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji.R
\name{ji}
\alias{ji}
\title{ji}
\usage{
ji(x)
}
\arguments{
\item{x, }{a numeric vector}
}
\value{
the product of all elements
}
\description{
Given a numeric vector \code{x}, generate the product of all elements
}
\examples{
x<-c(1,2,3)
ji(x)   ## result is 6

}
