% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawData.R
\name{rawData}
\alias{rawData}
\title{rawData}
\usage{
rawData(land, module, date, language = "")
}
\arguments{
\item{land, }{a data source or category}

\item{module, }{a data form which includes different information}

\item{date, }{special date when the entity existed}
}
\value{
a set rawData including info such as id, name, type and coordinates etc
}
\description{
Given \code{land}, data form \code{module} and \code{date}, generate the raw data set
}
\examples{
info<-rawData(land,"info",date)$description
         

}
