% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk1.R
\name{walk1}
\alias{walk1}
\title{walk}
\usage{
walk1(i, z)
}
\arguments{
\item{i, }{the first index}

\item{z, }{the coordinates list extracted from dataset}
}
\value{
the corresponding ggplot aesthetics including geo coordinates
}
\description{
Given index \code{i} and the coordinates list \code{z}, generate the corresponding ggplot aesthetics
}
\examples{
x<-ICU$new(land="world-2",date="1984-11")   ## New class
z<-x$guo$Countries$coordinates
ggplot()+walk1(1,z)  ## Western Sahara

}
