% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk2.R
\name{walk2}
\alias{walk2}
\title{walk2}
\usage{
walk2(i, j, z)
}
\arguments{
\item{i, }{the first index}

\item{j, }{the second index}

\item{z, }{the coordinates list extracted from dataset}
}
\value{
the corresponding ggplot aesthetics including geo coordinates
}
\description{
Given index \code{i}, \code{j} and the coordinates list \code{z}, generate the corresponding ggplot aesthetics
}
\examples{
x<-ICU$new(land="world-2",date="1984-11")   ## New class
z<-x$guo$Countries$coordinates
ggplot()+walk2(164,1,z)  ## the first area of Finland

}
