% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk3.R
\name{walk3}
\alias{walk3}
\title{walk3}
\usage{
walk3(i, j, k, z)
}
\arguments{
\item{i, }{the first index}

\item{j, }{the second index}

\item{k, }{the third index}

\item{z, }{the coordinates list extracted from dataset}
}
\value{
the corresponding ggplot aesthetics including geo coordinates
}
\description{
Given index \code{i}, \code{j}, \code{k} and the coordinates list \code{z}, generate the corresponding ggplot aesthetics
}
\examples{
x<-ICU$new(land="world-2",date="1984-11")   ## New class
z<-x$guo$Countries$coordinates
ggplot()+walk3(116,1,1,z)  ## the first area of Jordan

}
